[hail] create relatedness group
This groups IBD and pc_relate together. I will follow up with KING. I plan to
lift some of the documentation verbiage from PC-Relate up and to unify the
mathematical presentation for PC-Relate and KING.
