+++ hail/python/hail/methods/__init__.py
@@ -1,16 +1,20 @@
-from .family_methods import trio_matrix, mendel_errors, transmission_disequilibrium_test, de_novo
-from .impex import export_elasticsearch, export_gen, export_bgen, export_plink, export_vcf, \
-    import_locus_intervals, import_bed, import_fam, grep, import_bgen, import_gen, import_table, \
-    import_plink, read_matrix_table, read_table, get_vcf_metadata, import_vcf, import_gvcfs, \
-    import_vcfs, index_bgen, import_matrix_table
-from .statgen import skat, identity_by_descent, impute_sex, \
-    genetic_relatedness_matrix, realized_relationship_matrix, pca, \
-    hwe_normalized_pca, pc_relate, split_multi, filter_alleles, filter_alleles_hts, \
-    split_multi_hts, balding_nichols_model, ld_prune, row_correlation, ld_matrix, \
-    linear_mixed_model, linear_regression_rows, _linear_regression_rows_nd, logistic_regression_rows, poisson_regression_rows, \
-    linear_mixed_regression_rows, lambda_gc
+from .family_methods import (trio_matrix, mendel_errors,
+                             transmission_disequilibrium_test, de_novo)
+from .impex import (export_elasticsearch, export_gen, export_bgen, export_plink,
+                    export_vcf, import_locus_intervals, import_bed, import_fam, grep,
+                    import_bgen, import_gen, import_table, import_plink, read_matrix_table,
+                    read_table, get_vcf_metadata, import_vcf, import_gvcfs, import_vcfs,
+                    index_bgen, import_matrix_table)
+from .statgen import (skat, impute_sex, genetic_relatedness_matrix,
+                      realized_relationship_matrix, pca, hwe_normalized_pca, split_multi,
+                      filter_alleles, filter_alleles_hts, split_multi_hts, balding_nichols_model,
+                      ld_prune, row_correlation, ld_matrix, linear_mixed_model,
+                      linear_regression_rows, _linear_regression_rows_nd,
+                      logistic_regression_rows, poisson_regression_rows,
+                      linear_mixed_regression_rows, lambda_gc)
 from .qc import sample_qc, variant_qc, vep, concordance, nirvana, summarize_variants
 from .misc import rename_duplicates, maximal_independent_set, filter_intervals
+from .relatedness import identity_by_descent, pc_relate
 
 __all__ = ['trio_matrix',
            'linear_mixed_model',
