+++ hail/python/hail/methods/relatedness/__init__.py
@@ -1,0 +1,7 @@
+from .identity_by_descent import identity_by_descent
+from .pc_relate import pc_relate
+
+__all__ = [
+    'identity_by_descent',
+    'pc_relate'
+]
