+++ hail/python/test/hail/methods/relatedness/test_identity_by_descent.py
@@ -1,0 +1,68 @@
+import os
+import unittest
+from subprocess import DEVNULL, call as syscall
+
+import hail as hl
+import hail.utils as utils
+from ..helpers import (startTestHailContext, stopTestHailContext, get_dataset,
+                       skip_unless_spark_backend)
+
+setUpModule = startTestHailContext
+tearDownModule = stopTestHailContext
+
+
+@skip_unless_spark_backend()
+@unittest.skipIf('HAIL_TEST_SKIP_PLINK' in os.environ, 'Skipping tests requiring plink')
+def test_ibd(self):
+    dataset = get_dataset()
+
+    def plinkify(ds, min=None, max=None):
+        vcf = utils.new_temp_file(prefix="plink", extension="vcf")
+        plinkpath = utils.new_temp_file(prefix="plink")
+        hl.export_vcf(ds, vcf)
+        threshold_string = "{} {}".format("--min {}".format(min) if min else "",
+                                          "--max {}".format(max) if max else "")
+
+        plink_command = "plink --double-id --allow-extra-chr --vcf {} --genome full --out {} {}" \
+            .format(utils.uri_path(vcf),
+                    utils.uri_path(plinkpath),
+                    threshold_string)
+        result_file = utils.uri_path(plinkpath + ".genome")
+
+        syscall(plink_command, shell=True, stdout=DEVNULL, stderr=DEVNULL)
+
+        ### format of .genome file is:
+        # _, fid1, iid1, fid2, iid2, rt, ez, z0, z1, z2, pihat, phe,
+        # dst, ppc, ratio, ibs0, ibs1, ibs2, homhom, hethet (+ separated)
+
+        ### format of ibd is:
+        # i (iid1), j (iid2), ibd: {Z0, Z1, Z2, PI_HAT}, ibs0, ibs1, ibs2
+        results = {}
+        with open(result_file) as f:
+            f.readline()
+            for line in f:
+                row = line.strip().split()
+                results[(row[1], row[3])] = (list(map(float, row[6:10])),
+                                             list(map(int, row[14:17])))
+        return results
+
+    def compare(ds, min=None, max=None):
+        plink_results = plinkify(ds, min, max)
+        hail_results = hl.identity_by_descent(ds, min=min, max=max).collect()
+
+        for row in hail_results:
+            key = (row.i, row.j)
+            self.assertAlmostEqual(plink_results[key][0][0], row.ibd.Z0, places=4)
+            self.assertAlmostEqual(plink_results[key][0][1], row.ibd.Z1, places=4)
+            self.assertAlmostEqual(plink_results[key][0][2], row.ibd.Z2, places=4)
+            self.assertAlmostEqual(plink_results[key][0][3], row.ibd.PI_HAT, places=4)
+            self.assertEqual(plink_results[key][1][0], row.ibs0)
+            self.assertEqual(plink_results[key][1][1], row.ibs1)
+            self.assertEqual(plink_results[key][1][2], row.ibs2)
+
+    compare(dataset)
+    compare(dataset, min=0.0, max=1.0)
+    dataset = dataset.annotate_rows(dummy_maf=0.01)
+    hl.identity_by_descent(dataset, dataset['dummy_maf'], min=0.0, max=1.0)
+    hl.identity_by_descent(dataset, hl.float32(dataset['dummy_maf']), min=0.0, max=1.0)
+
